requirements.txt
requests==2.21.0
selenium==3.141.0
pyotp==2.6.0
Pillow==8.3.1
python-dotenv==0.18.0
